+++ src/contraction/fwd_ch_dd_expansion_policy.h
@@ -44,7 +44,7 @@ class fwd_ch_dd_expansion_policy : public expansion_policy
         virtual void
         get_xy(warthog::search_node*, int32_t& x, int32_t& y);
 
-        virtual uint32_t
+        virtual size_t
         mem()
         {
             return expansion_policy::mem() +
